+++ WebApplication4/Controllers/EventsController.cs
@@ -66,6 +66,8 @@ namespace DogsServer.Controllers
         [HttpPost]
         public IActionResult Post([FromBody]EventModel obj)
         {
+            if (!IsCurrentUserAdmin() && !IsCurrentUserMember())
+                return Forbid();
             var oneEvent = new Event
             {
                 EventId = obj.EventId,
@@ -85,6 +87,8 @@ namespace DogsServer.Controllers
         [HttpPut("{id}")]
         public IActionResult Put(int id, [FromBody]EventModel obj)
         {
+            if (!IsCurrentUserAdmin() && !IsCurrentUserMember())
+                return Forbid();
             var oldEvent = unitOfWork.EventRepository.GetById(id);
             oldEvent.City = obj.City;
             oldEvent.Date = obj.Date;
@@ -100,6 +104,8 @@ namespace DogsServer.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            if (!IsCurrentUserAdmin() && !IsCurrentUserMember())
+                return Forbid();
             unitOfWork.EventRepository.Delete(unitOfWork.EventRepository.GetById(id));
             unitOfWork.Commit();
             return new ObjectResult("Event deleted successfully!");
